# benchmarks: fix reported MB/sec values