--- a/storage/memoized_iterator_test.go
+++ b/storage/memoized_iterator_test.go
@@ -75,7 +75,7 @@ func BenchmarkMemoizedSeriesIterator(b *testing.B) {
 	// Simulate a 5 minute rate.
 	it := NewMemoizedIterator(newFakeSeriesIterator(int64(b.N), 30), 5*60)
 
-	b.SetBytes(int64(b.N * 16))
+	b.SetBytes(16)
 	b.ReportAllocs()
 	b.ResetTimer()